+++ src/lib/exception/SlackAuthenticationError.ts
@@ -1,0 +1,5 @@
+export default class SlackAuthenticationError extends Error {
+    constructor(extraMessage?: string) {
+        super(`Slack authentication error${extraMessage ? ` ${extraMessage}` : ''}`);
+    }
+}
